Added untested Proximity-Based Weighting Algorithm
This commit adds the code that calculates suspiciousness scores
according to the proximity-based weighting algorithm. I haven't tested
it yet or included all the documentation I want to - I'm in the middle
of traveling right now and don't have time to do so xD. I'll add the
missing tests and documentation, as well as fix any bugs I find in doing
so, in the next commit or few.