Tests are now passing
It turns out that the reason my tests were failing wasn't because
my program had a bug. Instead, the example from the paper I was
using didn't perform the full algorithm, but instead picked arbitrary
values for test weights. Had the paper used the whole algorithm instead,
it would have gotten the same results my program currently gets.

So the tests are now working, and I can get to work on the rest of the
program!