Ignore dummy app
To test the Historical Fault Localization tool, I've created a simple
Ruby program and written an extensive test suite for it. It is
version controlled on its own, so it there is no need to version it
with the Historical Fault Localization tool. In addition, the
Historical Fault Localization tool will manipulate the Ruby app's
repository using git, which would be problematic if the app was
included in the Historical Fault Localization repository.